apply workspace linting to all crates and related cleanups

+++ crates/processors/tests/js_processor_tests.rs
@@ -26,7 +26,7 @@ fn new_event() -> Event {
         trace_id: None,
         tags: None,
         checkpoint: None,
-        size_bytes: 20 as usize,
+        size_bytes: 20_usize,
         tx_end: true,
     }
 }
